Remove misleading information
The core benefit of a prepared statement is not the performance gain! 

It's the server side SQL injection escaping. If you use prepared statements it is impossible to have SQL injection vulnerabilities, because the database will take care of escaping for you.

If in anything it is premature optimization not to use them, because you are giving up immense security benefits for a slight performance improvement.